produtos

+++ backend/src/main/java/com/ecosun/dto/ProdutoRequest.java
@@ -1,11 +1,9 @@
 package com.ecosun.dto;
 
-import java.math.BigDecimal;
-
 public class ProdutoRequest {
     private String nome;
     private String descricao;
-    private BigDecimal preco;
+    private Double preco;
     private String foto;
     private Integer categoriaId;
 
@@ -17,8 +15,8 @@ public class ProdutoRequest {
     public String getDescricao() { return descricao; }
     public void setDescricao(String descricao) { this.descricao = descricao; }
 
-    public BigDecimal getPreco() { return preco; }
-    public void setPreco(BigDecimal preco) { this.preco = preco; }
+    public Double getPreco() { return preco; }
+    public void setPreco(Double preco) { this.preco = preco; }
 
     public String getFoto() { return foto; }
     public void setFoto(String foto) { this.foto = foto; }
